reposition adjusted to create next 3 Balls in unique y-axis pos

+++ core/src/com/ammonf/game/sprites/BadBall.java
@@ -4,7 +4,7 @@ package com.ammonf.game.sprites;
  * Created by Fritz on 5/14/2016.
  */
 public class BadBall extends Ball {
-    public BadBall(float x) {
+    public BadBall(int x) {
         super(x, "ball-bad1.png");
     }
